Move checkpointing utils into checkpoint_utils.py
