+++ fairseq/models/fconv_self_att.py
@@ -4,7 +4,6 @@
 # This source code is licensed under the license found in the LICENSE file in
 # the root directory of this source tree. An additional grant of patent rights
 # can be found in the PATENTS file in the same directory.
-#
 
 import math
 
@@ -12,11 +11,11 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
+from fairseq import checkpoint_utils
 from fairseq.modules import (
     DownsampledMultiHeadAttention, GradMultiply, LayerNorm,
     LearnedPositionalEmbedding, LinearizedConvolution,
 )
-from fairseq import utils
 
 from . import (
     FairseqEncoder, CompositeEncoder, FairseqDecoder, FairseqModel,
@@ -84,8 +83,7 @@ class FConvModelSelfAtt(FairseqModel):
         pretrained = eval(args.pretrained)
         if pretrained:
             print("| loading pretrained model")
-            trained_model = utils.load_ensemble_for_inference(
-                # not actually for inference, but loads pretrained model parameters
+            trained_model = checkpoint_utils.load_model_ensemble(
                 filenames=[args.pretrained_checkpoint],
                 task=task,
             )[0][0]
